% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEBDF2Freq.R
\name{BEBDF2Freq}
\alias{BEBDF2Freq}
\title{Create a frequency table from the parsed Bayes Empiracle Bayes data frame}
\usage{
BEBDF2Freq(df_beb, only_signif = NULL, max_freq = NULL)
}
\arguments{
\item{df_beb}{data frame from parse_BEB() function}

\item{only_signif}{Set to TRUE to select only significant sites}

\item{max_freq}{Keep genes who have n-sites under selection up to or equal to this}
}
\description{
Creates a frequency table based on the parsed output from the parse_BEB() function
}
\examples{
BEBDF2Freq(df_beb = df.beb, only_signif = TRUE, max_freq = 100)
}
\keyword{frequency}
\keyword{table}
