% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylip_length.R
\name{clean_seqret}
\alias{clean_seqret}
\title{Trim phylip files to make them a multiple of three}
\usage{
clean_seqret(infile, headers, file_dir = NULL)
}
\arguments{
\item{infile}{File path to phylip file}

\item{headers}{A vector of the sequence headers}

\item{file_dir}{An output file directory (doesn't need to exist)}
}
\description{
Checks if the sequence in a phylip file is a multiple of three. If it is not, it trims the minimum number of bases needed to make it so.
}
\examples{
clean_seqret(infile = "/path/to/phylip.phy", headers = c("sample1", "sample2", "sample3"), file_dir = "/path/to/output_directory")
}
\keyword{clean,}
\keyword{phylip,}
\keyword{seqret}
\keyword{trim,}
