% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_bcm.R
\name{.lrt_bcm}
\alias{.lrt_bcm}
\title{Calculate LRT statistic for when both models are Branch/clade/branch-site models}
\usage{
.lrt_bcm(df, mdl)
}
\arguments{
\item{df}{Long-format dataframe to parse for model information}

\item{mdl}{Branch/clade/branch-site models to compare}
}
\description{
Builds a table from the np/lnL data and runs an LRT analysis between two models.
}
\examples{
.lrt_bcm(df = df, mdl = c("CmC", "M2a_Rel"))
}
\keyword{internal}
