% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_hnh.R
\name{.lrt_hnh}
\alias{.lrt_hnh}
\title{Calculate LRT statistic for models from each group (nsm/bcm)}
\usage{
.lrt_hnh(df, mdl)
}
\arguments{
\item{df}{Long-format dataframe to parse}

\item{mdl}{Vector of models for LRT statistic}
}
\description{
Calcualtes LRT statistic between two models, where one belongs to Branch/Clade/Branch-site and the other
to Null/Site. Name derived from "half-n-half".
}
\examples{
.lrt_hnh(df = df.long, mdl = c("ModelA1", "M1a"))
}
\keyword{internal}
