% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_nsm.R
\name{.lrt_nsm}
\alias{.lrt_nsm}
\title{Calculate LRT statistic for Null/Site models}
\usage{
.lrt_nsm(df, mdl)
}
\arguments{
\item{df}{Long form dataframe}

\item{mdl}{Vector of two null/site models for comparison}
}
\description{
Builds a table from the np/lnL data and runs an LRT analysis between two models.
}
\examples{
.lrt_nsm(df = df.long, mdl = c("M1a", "M2a"))
}
\keyword{internal}
