% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_m2a_M8.R
\name{.parse_m2a_M8}
\alias{.parse_m2a_M8}
\title{Parse BEB output from M2a and M8 codeml output}
\usage{
.parse_m2a_M8(lines, c)
}
\arguments{
\item{lines}{List of codeml output files}

\item{c}{number of cores}
}
\description{
Internal function for parsing the BEB output from M2a and M8 output files.
Parsing works by using anchors specific to these models
}
\keyword{internal}
