% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ModelA.R
\name{.parse_ModelA}
\alias{.parse_ModelA}
\title{Parse BEB output from ModelA codeml output}
\usage{
.parse_ModelA(lines, c)
}
\arguments{
\item{lines}{List of codeml output files}

\item{c}{number of cores}
}
\description{
Extracts the BEB information from ModelA codeml output files.
Uses anchors specific to Model A to parse data.
}
\keyword{internal}
