% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_np_lnL.R
\name{.parse_np_lnL}
\alias{.parse_np_lnL}
\title{Parse lnL and np data from codeml output files}
\usage{
.parse_np_lnL(vec.list)
}
\arguments{
\item{vec.list}{list of vectors from readr::read_lines() command (corresponds to single codeml.log file)}
}
\description{
Internal function for the lrt_calculation() function.
}
\keyword{internal}
