% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBEB.R
\name{getBEB}
\alias{getBEB}
\title{Extract Bayes Empiracle Bayes information from codeml output}
\usage{
getBEB(dir_path, models, sig = NULL, cores = 1, ext = ".out")
}
\arguments{
\item{dir_path}{Directory path to codeml_parallel output directory}

\item{models}{Vector of models to run. Valid models include `M2a`, `M8` and `ModelA`}

\item{sig}{P-value threshold to filter BEB output by e.g. `sig = 0.05`}

\item{cores}{Number of cores to parallelise over}

\item{ext}{Extension of CODEML output files}
}
\description{
Extracts all sites reported in the Bayes Empiracle Bayes section of the codeml output as a dataframe object
}
\examples{
getBEB(dir_path = "path/to/parent_out", models = c("M2a", "M8", "ModelA"), sig = 0.1, cores = 4, ext = ".customOut")
}
\keyword{Bayes,}
\keyword{empiracle,}
\keyword{parse}
