% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrthologHeaders.R
\name{getOrthologHeaders}
\alias{getOrthologHeaders}
\title{Returns a key-value table of the gene symbols (filename) and fasta headers corresponding to those genes}
\usage{
getOrthologHeaders(crbb_path, crbb_ext, sample_separator, id_pct, aln_pct)
}
\arguments{
\item{crbb_path}{Path to the directory containing CRBB files}

\item{crbb_ext}{Extension of the CRBB files}

\item{sample_separator}{Delimiter between sample IDs in the filename}

\item{id_pct}{Percentage identity between hits}

\item{aln_pct}{Proportion of query/target that needs to be accounted for in BLAST alignment portion}
}
\description{
Given a path to a directory containing conditional reciprocal best blast (CRBB) files, this function will
parse the data and find sequences that are considered a CRBB across multiple comparisons. This function
relies on one sample being the reference for the remaining comparisons to be joined to.
}
\details{
CRBB files need to have the filenaming structure as follows; sampleA_sampleB.ext. Both sample identifiers need
to be present in the file name, separated by some kind of delimeter. The delimiter does not matter, as you can
specify this at the function call.
}
\examples{
getOrthologHeaders(crbb_path = "path/to/directory",
crbb_ext = ".tsv",
sample_separator = "__",
id_pct = 95,
aln_pct = 90)
}
\keyword{helper}
