% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPW.R
\name{getPW}
\alias{getPW}
\title{Parse the `p` and `w` fields from codeml output files}
\usage{
getPW(dir_path, models, ext = ".out")
}
\arguments{
\item{dir_path}{Path to codeml_parallel output directory}

\item{models}{Vector of models to extract p/w information from}

\item{ext}{Extension of output files}
}
\description{
Extracts the p and w lines from codeml output and stores it in a list/data frame object
}
\keyword{parse}
