% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goFisher.R
\name{goFisher}
\alias{goFisher}
\title{Significance testing of GO terms for grouped data}
\usage{
goFisher(gene_2_go, gene_2_group, pAdj_method, p_cutoff)
}
\arguments{
\item{gene_2_go}{Dataframe that contains two columns. Column 1 is gene ID and column 2 is GO terms associated with that gene ID.}

\item{gene_2_group}{Long form dataframe mapping which group a gene belongs to. Genes can appear multiple times belonging to different groups.
Column 1 should be the gene ID and column two is the grouping variable.}

\item{pAdj_method}{P-value adjustment method. Takes any value from the `p.adjust()`` function}

\item{p_cutoff}{P-value threshold that values must be less than or equal to}
}
\description{
Conducts a Fisher's exact test on a 2x2 contigency table of presence/absence of GO terms for grouped data.
}
\keyword{helper}
