% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_fasta.R
\name{import_fasta}
\alias{import_fasta}
\title{Accessory function to read in fasta files to R}
\usage{
import_fasta(cds_file_path, ext_cds, custom_names)
}
\arguments{
\item{cds_file_path}{Path to the fasta files directory}

\item{ext_cds}{Extension of fasta files}

\item{custom_names}{Names for the fasta list elements. NEED to be in the same order as the files are read in.}
}
\description{
Imports fasta files to R using the seqinR package.
}
\examples{
import_fasta(cds_file_path = "/path/to/directory", ext_cds = ".phy", custom_names = c("sample_1", "sample_2", "sample_3"))
}
\keyword{helper}
