% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_keyValue.R
\name{import_keyValue}
\alias{import_keyValue}
\title{Accessory function to import key-value pair file}
\usage{
import_keyValue(kv_file_path)
}
\arguments{
\item{kv_file_path}{Path to the key-value pair file}
}
\description{
Imports the key-value pair file for use with write_fasta()
}
\examples{
import_keyValue(kv_file_path = "path/to/kv.tsv")
}
\keyword{helper}
