% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_stops.R
\name{internal_stops}
\alias{internal_stops}
\title{Identify gene alignments that have more than one stop codon in the sequence}
\usage{
internal_stops(dir_path, model_string, tree_string, nSites = NULL)
}
\arguments{
\item{dir_path}{Directory path to codeml_parallel output directory.}

\item{model_string}{Model to import information for. This is arbitrary as all comparisons will use the same alignment.}

\item{tree_string}{If there are more than one tree, specify which tree to read in. Again, this is arbitrary as all gene-models will have the same alignment}

\item{nSites}{Integer threshold. Extract genes that have this many or greater stop codons in their alignment.}
}
\description{
Parses the codeml output for the number of columns converted to '???' in the sequence alignment for each gene
}
\examples{
internal_stops(dir_path = "path/to/codeml_out", model_string = "Model2Selection", tree_string = "brownForeground", nSites = 2)
}
\keyword{column,}
\keyword{converted}
\keyword{stop,}
