% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSequential.R
\name{isSequential}
\alias{isSequential}
\title{Identify sites under selection that are in close proximity to other sites under selection}
\usage{
isSequential(beb_long, distance_threshold)
}
\arguments{
\item{beb_long}{Long form BEB table from getBEB() function}

\item{distance_threshold}{Report sites that are within this proximity to other reported sites under selection}
}
\description{
This function checks the distance between sites under selection and reports those that
fall within a specified distance of each other. For example, if you have two sites reported as
under selection, but they are the two amino acids next to each other, both sites would be reported
if the distance_threshold was set to '1'.
}
\examples{
isSequential(beb_long = df.beb, distance_threshold = 1)
}
\keyword{helper}
