% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt_calculation.R
\name{lrt_statistic}
\alias{lrt_statistic}
\title{Description}
\usage{
lrt_statistic(dir_path, lst_models, lst_comparisons)
}
\arguments{
\item{dir_path}{Path to codeml_parallel output directory.}

\item{lst_models}{A list of models to use in the LRT calculations}

\item{lst_comparisons}{Combinations/comparisons of the above models}
}
\description{
General run down of what it does
}
\examples{
lrt_statistic(dir_path = "path/to/codeml_out", lst_models = c("Model1Neutral", "Model2Selection"), lst_comparisons = c("Model1Neutral", "Model2Selection"))
}
\keyword{LRT}
\keyword{Likelihood}
\keyword{ratio}
\keyword{test,}
