% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrtStatistic.R
\name{lrtStatistic}
\alias{lrtStatistic}
\title{Conduct Likelihood-Ratio-Test (LRT) between two evolutionary models}
\usage{
lrtStatistic(dir_path, comp_list, ext = ".out")
}
\arguments{
\item{dir_path}{Path to `parallel_codeml` output directory.}

\item{comp_list}{Model combinations as a list of vectors}

\item{ext}{Extension of CODEML output files. Defaults to ".out"}
}
\description{
This function takes the path to the parallel_codeml output directory and a list of comparisons as input.
Based on these variables, it will go and run LRTs for each comparison, generating a list of data frames
containing all parameters used to run the statistic, as well as the result.
}
