% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_BEB.R
\name{parse_BEB}
\alias{parse_BEB}
\title{Extract Bayes Empiracle Bayes information from codeml output}
\usage{
parse_BEB(dir_path, out_path = NULL)
}
\arguments{
\item{dir_path}{Directory path to codeml_parallel output directory}

\item{out_path}{Directory of where to write data frame object generated by parse_BEB()}
}
\description{
Extracts all sites reported in the Bayes Empiracle Bayes section of the codeml output as a dataframe object
}
\examples{
parse_BEB(dir_path = "path/to/parent_out", out_path = "path/to/output/name.tsv")
}
\keyword{Bayes,}
\keyword{empiracle,}
\keyword{parse}
