% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_PW.R
\name{parse_PW}
\alias{parse_PW}
\title{Parse the `p` and `w` fields from codeml output files}
\usage{
parse_PW(dir_path, df_out = NULL)
}
\arguments{
\item{dir_path}{Path to codeml_parallel output directory.}

\item{df_out}{Set at TRUE to output a long form data frame. Default is a list object; one element per gene-tree combination.}
}
\description{
Extracts the p and w lines from codeml output and stores it in a list/data frame object
}
\examples{
parse_PW(parse_PW = "path/to/codeml_out", df_out = TRUE)
}
\keyword{parse}
