% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_multiFasta.R
\name{write_multiFasta}
\alias{write_multiFasta}
\title{Writes a multi-fasta file by gene}
\usage{
write_multiFasta(df_kevValue, lst_fastas, out_dir)
}
\arguments{
\item{df_kevValue}{Key-value data frame}

\item{lst_fastas}{List of fasta files}

\item{out_dir}{Output directory path}
}
\description{
Writes a multi-fasta by gene, where each sequence in a file corresponds to the gene sequence from a different sample.
}
\examples{
write_multiFasta(df_kevValue = df.kv, lst_fastas = lst.fa, out_dir = "path/to/output/dir")
}
\keyword{by-gene}
\keyword{fasta,}
\keyword{write}
