% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFasta.R
\name{writeFasta}
\alias{writeFasta}
\title{Writes a multi-fasta file by gene for n-species}
\usage{
writeFasta(orthologs, fasta_dir, pep_ext, nuc_ext, pep_out, nuc_out,
  stop_codon = ".", write_file = TRUE)
}
\arguments{
\item{orthologs}{Output object from `getOrthologueHeaders()` function}

\item{fasta_dir}{Directory where fasta files (peptide and nucleotide) are located for each sample in table}

\item{pep_ext}{Extension of peptide files}

\item{nuc_ext}{Extension of nucleotide files}

\item{pep_out}{Path to output directory for peptide files}

\item{nuc_out}{Path to output directory for nucleotide files}

\item{stop_codon}{Character that is used to specify what a stop codon is in your sequence. Defaul = "."}

\item{write_file}{Set as `FALSE`` if you do not want to write fasta files. Default is TRUE}
}
\description{
This function creates fasta files for each gene sequence from the orthologue table (`getOrthologueHeaders``).
}
\examples{
writeFasta(orthologs = listObj,
fasta_dir = "path/to/fasta/directory",
nuc_ext = ".fna",
pep_ext = ".pep",
nuc_out = "path/to/nuc_out/dir",
pep_out = "/path/to/pep_out/dir",
stop_codon = "*",
write_file = FALSE)
}
\keyword{by-gene}
\keyword{fasta,}
\keyword{write}
