% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFreq.R
\name{getFreq}
\alias{getFreq}
\title{Create a frequency table from the parsed Bayes Empiracle Bayes data}
\usage{
getFreq(df_long, significant = NULL, site_as_prop = NULL)
}
\arguments{
\item{df_long}{Long form data frame from the getBEB() function}

\item{site_as_prop}{Return frequencies values as a proportion of the genes length instead of simple counts}

\item{only_signif}{Set to TRUE to select only significant sites}
}
\description{
Creates a frequency table based on the parsed output from the getBEB() function in the cmlHelpR package.
}
\examples{
getFreq(df_long = df.beb, significant = TRUE, site_as_prop = TRUE)
}
\keyword{frequency}
\keyword{table}
